move nearest 10 vp script into old script
