+++ rt_segment_speeds/scripts/vp_around_stops.py
